refactor: mover tendencias como subtab del balance muscular y eliminar tab independiente
- Agregado nuevo subtab 'Tendencias' al Balance Muscular con icono Brain
- Movido todo el contenido del TrendsTab al nuevo subtab manteniendo funcionalidad
- Eliminado tab independiente de 'Tendencias' del dashboard
- Actualizado tipos, constantes y componentes del dashboard
- Eliminado archivo trends-tab.tsx completamente
- Corregido tipos y referencias en dashboard-page.tsx y dashboard-modal.tsx
