+++ src/components/dashboard/dashboard-modal.tsx
@@ -9,8 +9,7 @@ import {
   DashboardEmptyState,
   ExercisesTab,
   HistoryTab,
-  PredictionsTab,
-  TrendsTab
+  PredictionsTab
 } from './components';
 import { DEFAULT_DASHBOARD_TAB } from './constants';
 import { useDashboardData } from './hooks';
@@ -127,8 +126,6 @@ export const Dashboard: React.FC<DashboardProps> = ({ onClose }) => {
                 switch (activeTab) {
                   case 'balance':
                     return <BalanceTab records={workoutRecords} />;
-                  case 'trends':
-                    return <TrendsTab records={workoutRecords} />;
                   case 'history':
                     return <HistoryTab records={workoutRecords} />;
                   case 'exercises':
