+++ src/components/dashboard/types.ts
@@ -3,7 +3,7 @@
  * Tabs disponibles en el dashboard
  * Ahora incluye history y exercises junto con los tabs existentes
  */
-export type DashboardTab = 'balance' | 'trends' | 'advanced' | 'predictions' | 'history' | 'exercises';
+export type DashboardTab = 'balance' | 'advanced' | 'predictions' | 'history' | 'exercises';
 
 export interface DashboardTabConfig {
   id: DashboardTab;
